Overland Flow with Unit Test
This code tests to see that the Q goes steady at three different locations along x at the appropriate time scale. The time to test each location is determined by the steady state drainage response timescale. If the Q is not in steady state when it is supposed to be at each location, the code will stop running and send the message 'code sucks.'